+++ src/components/custom/item/ItemPop.vue
@@ -6,15 +6,15 @@ import {
 import {
   OpenInNewFilled
 } from '@vicons/material'
-import XivFARImage from './XivFARImage.vue'
 import ItemSpan from './ItemSpan.vue'
 import ItemRemark from './ItemRemark.vue'
+import XivFARImage from '../general/XivFARImage.vue'
+import LocationSpan from '../map/LocationSpan.vue'
 import { getItemInfo, type ItemInfo } from '@/tools/item'
 import type { UserConfigModel } from '@/models/user-config'
 import XivAttributes from '@/assets/data/xiv-attributes.json'
 import { XivItemRemarks, XivJobs, type XivJob } from '@/assets/data'
 import type EorzeaTime from '@/tools/eorzea-time'
-import LocationSpan from './LocationSpan.vue'
 
 const t = inject<(text: string, ...args: any[]) => string>('t') ?? (() => { return '' })
 const isMobile = inject<Ref<boolean>>('isMobile') ?? ref(false)
