+++ src/components/main/StatisticsPanel.vue
@@ -3,11 +3,11 @@ import { computed, inject, ref, type Ref } from 'vue'
 import {
   NButton, NCollapse, NCollapseItem
 } from 'naive-ui'
-import FoldableCard from '../custom-controls/FoldableCard.vue'
-import GroupBox from '../custom-controls/GroupBox.vue'
-import ItemButton from '../custom-controls/ItemButton.vue'
-import ItemList from '../custom-controls/ItemList.vue'
-import TomeScriptButton from '../custom-controls/TomeScriptButton.vue'
+import FoldableCard from '../templates/FoldableCard.vue'
+import GroupBox from '../templates/GroupBox.vue'
+import ItemButton from '../custom/item/ItemButton.vue'
+import ItemList from '../custom/item/ItemList.vue'
+import TomeScriptButton from '../custom/other/TomeScriptButton.vue'
 import ModalCraftStatements from '../modals/ModalCraftStatements.vue'
 import ModalProStatements from '../modals/ModalProStatements.vue'
 import ModalCostAndBenefit from '../modals/ModalCostAndBenefit.vue'
