+++ src/components/templates/FoldableCard.vue
@@ -3,9 +3,11 @@ import { ref, type Ref, shallowRef, inject } from 'vue'
 import {
   NButton, NCard, NIcon
 } from 'naive-ui'
+import {
+  KeyboardArrowUpRound, KeyboardArrowDownRound
+} from '@vicons/material'
 import { useStore } from '@/store/index'
 import { type UserConfigModel } from '@/models/user-config'
-import { KeyboardArrowUpRound, KeyboardArrowDownRound } from '@vicons/material'
 
 const store = useStore()
