+++ src/views/FoodAndTincPage.vue
@@ -3,11 +3,11 @@ import { computed, inject, ref, watch, type Ref } from 'vue'
 import {
   FastfoodOutlined
 } from '@vicons/material'
-import { useNbbCal } from '@/tools/use-nbb-cal'
-import RouterCard from '@/components/subs/RouterCard.vue'
+import RouterCard from '@/components/custom/general/RouterCard.vue'
 import ItemSelectionPanel from '@/components/ft-helper/ItemSelectionPanel.vue'
 import StatisticsPanel from '@/components/ft-helper/StatisticsPanel.vue'
 import { useStore } from '@/store'
+import { useNbbCal } from '@/tools/use-nbb-cal'
 import type { UserConfigModel } from '@/models/user-config'
 
 const t = inject<(text: string, ...args: any[]) => string>('t') ?? (() => { return '' })
